pyproject.toml
[tool.poetry]
name = "resume-project"
version = "0.1.0"
description = ""
authors = ["Aleksey525 <alex_tolchin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.11.7"}
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
